requirements.txt
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<3.0.0
scikit-learn>=1.0.0,<2.0.0
tensorflow>=2.8.0,<3.0.0
torch>=1.12.0,<3.0.0
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.0.0,<6.0.0
networkx>=2.8.0,<4.0.0
scipy>=1.8.0,<2.0.0
joblib>=1.1.0,<2.0.0
streamlit>=1.28.0,<2.0.0
streamlit-option-menu>=0.3.0,<1.0.0
streamlit-aggrid>=0.3.0,<1.0.0
python-dotenv>=0.19.0,<2.0.0
requests>=2.28.0,<3.0.0
cryptography>=37.0.0,<42.0.0
paramiko>=2.11.0,<4.0.0
python-nmap>=0.7.1,<1.0.0
pyshark>=0.6.0,<1.0.0
scapy>=2.4.0,<3.0.0
pytorch-lightning>=1.8.0,<3.0.0
optuna>=3.0.0,<4.0.0
shap>=0.40.0,<1.0.0
lime>=0.2.0.1,<1.0.0
imbalanced-learn>=0.9.0,<1.0.0
xgboost>=1.6.0,<2.0.0
lightgbm>=3.3.0,<5.0.0
catboost>=1.1.0,<2.0.0
psutil>=5.8.0,<6.0.0
pyyaml>=6.0,<7.0 